+++ src/dynamics/multibody_joint.rs
@@ -67,12 +67,12 @@ impl RawMultibodyJointSet {
     }
 
     /// Return the lower limit along the given joint axis.
-    pub fn jointLimitsMin(&self, handle: FlatHandle, axis: RawJointAxis) -> f32 {
+    pub fn jointLimitsMin(&self, handle: FlatHandle, axis: RawJointAxis) -> f64 {
         self.map(handle, |j| j.data.limits[axis as usize].min)
     }
 
     /// If this is a prismatic joint, returns its upper limit.
-    pub fn jointLimitsMax(&self, handle: FlatHandle, axis: RawJointAxis) -> f32 {
+    pub fn jointLimitsMax(&self, handle: FlatHandle, axis: RawJointAxis) -> f64 {
         self.map(handle, |j| j.data.limits[axis as usize].max)
     }
 
@@ -92,10 +92,10 @@ impl RawMultibodyJointSet {
     pub fn jointConfigureBallMotorVelocity(
         &mut self,
         handle: FlatHandle,
-        vx: f32,
-        vy: f32,
-        vz: f32,
-        factor: f32,
+        vx: f64,
+        vy: f64,
+        vz: f64,
+        factor: f64,
     ) {
         let targetVel = Vector3::new(vx, vy, vz);
 
@@ -109,12 +109,12 @@ impl RawMultibodyJointSet {
     pub fn jointConfigureBallMotorPosition(
         &mut self,
         handle: FlatHandle,
-        qw: f32,
-        qx: f32,
-        qy: f32,
-        qz: f32,
-        stiffness: f32,
-        damping: f32,
+        qw: f64,
+        qx: f64,
+        qy: f64,
+        qz: f64,
+        stiffness: f64,
+        damping: f64,
     ) {
         let quat = Quaternion::new(qw, qx, qy, qz);
 
@@ -132,15 +132,15 @@ impl RawMultibodyJointSet {
     pub fn jointConfigureBallMotor(
         &mut self,
         handle: FlatHandle,
-        qw: f32,
-        qx: f32,
-        qy: f32,
-        qz: f32,
-        vx: f32,
-        vy: f32,
-        vz: f32,
-        stiffness: f32,
-        damping: f32,
+        qw: f64,
+        qx: f64,
+        qy: f64,
+        qz: f64,
+        vx: f64,
+        vy: f64,
+        vz: f64,
+        stiffness: f64,
+        damping: f64,
     ) {
         let quat = Quaternion::new(qw, qx, qy, qz);
         let vel = Vector3::new(vx, vy, vz);
@@ -160,8 +160,8 @@ impl RawMultibodyJointSet {
     //     &mut self,
     //     handle: FlatHandle,
     //     axis: RawJointAxis,
-    //     targetVel: f32,
-    //     factor: f32,
+    //     targetVel: f64,
+    //     factor: f64,
     // ) {
     //     self.jointConfigureMotor(handle, axis, 0.0, targetVel, 0.0, factor)
     // }
@@ -170,9 +170,9 @@ impl RawMultibodyJointSet {
     //     &mut self,
     //     handle: FlatHandle,
     //     axis: RawJointAxis,
-    //     targetPos: f32,
-    //     stiffness: f32,
-    //     damping: f32,
+    //     targetPos: f64,
+    //     stiffness: f64,
+    //     damping: f64,
     // ) {
     //     self.jointConfigureMotor(handle, axis, targetPos, 0.0, stiffness, damping)
     // }
@@ -181,10 +181,10 @@ impl RawMultibodyJointSet {
     //     &mut self,
     //     handle: FlatHandle,
     //     axis: RawJointAxis,
-    //     targetPos: f32,
-    //     targetVel: f32,
-    //     stiffness: f32,
-    //     damping: f32,
+    //     targetPos: f64,
+    //     targetVel: f64,
+    //     stiffness: f64,
+    //     damping: f64,
     // ) {
     //     self.map_mut(handle, |j| {
     //         j.data.motors[axis as usize].target_pos = targetPos;
